+++ app/src/main/java/in/hridayan/ashell/utils/app/CrashHandler.java
@@ -1,4 +1,4 @@
-package in.hridayan.ashell.utils;
+package in.hridayan.ashell.utils.app;
 
 import android.content.Context;
 import android.content.Intent;
