+++ app/src/main/java/in/hridayan/ashell/utils/app/updater/FetchLatestVersionCode.java
@@ -1,33 +1,52 @@
-package in.hridayan.ashell.utils;
+package in.hridayan.ashell.utils.app.updater;
 
 import android.content.Context;
-import android.os.AsyncTask;
+import android.os.Handler;
+import android.os.Looper;
 import in.hridayan.ashell.config.Const;
 import in.hridayan.ashell.config.Preferences;
+import in.hridayan.ashell.utils.DeviceUtils;
 import in.hridayan.ashell.utils.DeviceUtils.FetchLatestVersionCodeCallback;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
-public class FetchLatestVersionCode extends AsyncTask<String, Void, Integer> {
-  private Context context;
-  private FetchLatestVersionCodeCallback callback;
+public class FetchLatestVersionCode {
+  private final Context context;
+  private final FetchLatestVersionCodeCallback callback;
+  private final ExecutorService executor;
+  private final Handler handler;
 
   public FetchLatestVersionCode(Context context, FetchLatestVersionCodeCallback callback) {
     this.context = context;
     this.callback = callback;
+    this.executor = Executors.newSingleThreadExecutor();
+    this.handler = new Handler(Looper.getMainLooper());
   }
 
-  @Override
-  protected Integer doInBackground(String... params) {
+  public void execute(String url) {
+    executor.execute(
+        () -> {
+          int result = fetchVersionCode(url);
+          handler.post(
+              () -> {
+                if (callback != null) {
+                  callback.onResult(result);
+                }
+              });
+        });
+  }
+
+  private int fetchVersionCode(String urlString) {
     StringBuilder result = new StringBuilder();
     try {
-      URL url = new URL(params[0]);
+      URL url = new URL(urlString);
       HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-      try {
-        BufferedReader reader =
-            new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
+      try (BufferedReader reader =
+          new BufferedReader(new InputStreamReader(urlConnection.getInputStream()))) {
         String line;
         while ((line = reader.readLine()) != null) {
           result.append(line).append("\n");
@@ -46,13 +65,6 @@ public class FetchLatestVersionCode extends AsyncTask<String, Void, Integer> {
       return Const.UPDATE_AVAILABLE; // Update available
     } else {
       return Const.UPDATE_NOT_AVAILABLE; // No update available
-    }
-  }
-
-  @Override
-  protected void onPostExecute(Integer result) {
-    if (callback != null) {
-      callback.onResult(result);
     }
   }
 }
